+++ src/main/kotlin/com/sexysisters/tojserverv2/common/util/api/oauth/exception/OtherServerForbiddenException.kt
@@ -1,4 +1,4 @@
-package com.sexysisters.tojserverv2.common.util.api.exception
+package com.sexysisters.tojserverv2.common.util.api.oauth.exception
 
 import com.sexysisters.tojserverv2.common.exception.BaseException
